Rename "imports" to tprek
