+++ tests/test_external_services/test_tprek/test_tprek_api_client.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from tilavarauspalvelu.utils.importers.tprek_api_client import TprekAPIClient
+from tilavarauspalvelu.utils.tprek.tprek_api_client import TprekAPIClient
 from utils.date_utils import local_datetime
 
 from tests.helpers import patch_method
