+++ tilavarauspalvelu/admin/unit/admin.py
@@ -11,7 +11,7 @@ from modeltranslation.admin import TranslationAdmin
 
 from tilavarauspalvelu.admin.location.admin import LocationInline
 from tilavarauspalvelu.models import Unit
-from tilavarauspalvelu.utils.importers.tprek_unit_importer import TprekUnitImporter
+from tilavarauspalvelu.utils.tprek.tprek_unit_importer import TprekUnitImporter
 from utils.sentry import SentryLogger
 
 if TYPE_CHECKING:
